Add ResponseEntity to UserController

+++ src/main/java/pl/romanek/blog/controller/UserController.java
@@ -1,14 +1,18 @@
 package pl.romanek.blog.controller;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
 import pl.romanek.blog.entity.User;
 import pl.romanek.blog.service.UserService;
 
@@ -24,22 +28,30 @@ public class UserController {
     }
 
     @GetMapping("/all")
-    public List<User> getAllUsers() {
-        return userService.findAllUsers();
+    public ResponseEntity<List<User>> getAllUsers() {
+        List<User> users = new ArrayList<>(userService.findAllUsers());
+        return users.isEmpty() ? ResponseEntity.notFound().build() : ResponseEntity.ok(users);
     }
 
     @GetMapping("/{id}")
-    public User getUser(@PathVariable("id") Integer id) {
-        return userService.findUserById(id);
+    public ResponseEntity<User> getUser(@PathVariable("id") Integer id) {
+        Optional<User> user = userService.findUserById(id);
+        return user.isPresent() ? ResponseEntity.ok(user.get()) : ResponseEntity.notFound().build();
     }
 
     @PostMapping("/add")
-    public void addUser(User user) {
+    public ResponseEntity<User> addUser(User user) {
         userService.addUser(user);
+        return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
     @DeleteMapping("/delete/{id}")
-    public void deleteUser(@PathVariable("id") Integer id) {
-        userService.deleteUser(id);
+    public ResponseEntity<User> deleteUser(@PathVariable("id") Integer id) {
+        Optional<User> user = userService.findUserById(id);
+        if (user.isPresent()) {
+            userService.deleteUser(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 }
